refactor: Seperate features in Player file + functions
