+++ src/lib/components/Scene.svelte
@@ -4,7 +4,7 @@
   import { CollisionGroups } from "@threlte/rapier";
   import { spring } from "svelte/motion";
   import { Mesh, Vector3 } from "three";
-  import Player from "./Player.svelte";
+  import Player from "./player/Player.svelte";
   import SplashWall from "./map/SplashWall.svelte";
   import ObstacleObjects from "./map/ObstacleObjects.svelte";
   import GenericPlaform from "./map/GenericPlatform.svelte";
